latest changes

+++ src/SingleSchedule/SingleSchedule.jsx
@@ -1,7 +1,5 @@
 import React, { Fragment } from 'react';
 import moment from 'moment';
-import Arrow from '@material-ui/icons/ArrowRightAlt';
-
 
 class SingleSchedule extends React.Component {
 
@@ -15,22 +13,26 @@ class SingleSchedule extends React.Component {
     render() {
         if(this.props.prev === 1){
             var additional = null;
-        }else {
-          
-          additional = 
-        <td><button className="Add" onClick={() => this.props.deleteItem(this.props.startTime)}>Delete</button> 
-        </td>;
+       }else if (this.props.prev !== 1 && this.props.style !== undefined) {
+            additional = 
+            <td><button className="Add" onClick={() => this.props.deleteItem(this.props.id)}>Delete</button> 
+            </td>;
+        }else{
+            additional = 
+            <td><button className="Add" onClick={() => this.props.deleteItem(this.props.startTime)}>Delete</button> 
+            </td>;
         }
+  
 return (
     <Fragment>
-        <tr className = {this.props.style}>
+        <tr>
         <td></td>
-        <td className="collapsing">
+        <td className="collapsing" className = {this.props.style}>
         <input type="checkbox"/> <label></label>
         </td>
-        <td>{this.props.startTime}</td>
-        <td>{this.props.title}</td>
-        <td>{moment.duration(this.props.duration)._data.minutes}:{this.timeFormat()}</td>
+        <td className = {this.props.style}>{this.props.startTime}</td>
+        <td className = {this.props.style}>{this.props.title}</td>
+        <td className = {this.props.style}>{moment.duration(this.props.duration)._data.minutes}:{this.timeFormat()}</td>
         {additional}
     </tr>
     </Fragment>
